Rename button with checklist warning
