+++ assets/src/edit-story/components/header/buttons/buttonWithChecklistWarning.js
